requirements.txt
tensorflow>=2.11.0
streamlit>=1.25.0
numpy>=1.23.0
Pillow>=9.3.0
requests>=2.28.0
